uint64 issue

+++ Alpha/classes/@rise_sym/commit.m
@@ -6,7 +6,7 @@ end
 
 [eqtn,obj]=recompose(obj);
 if ~isKey(rise_sym_main_map.fid,eqtn);
-    Count=rise_sym_main_map.fid.Count+1;
+    Count=double(rise_sym_main_map.fid.Count)+1;
     ref=['ref_',rise_sym_main_map.tag,'_',sprintf('%.0f',Count),'_']; % adding an underscore latter so that we can use strrep, which should be faster, rather than regexp
     obj.ref=ref; % handle will push this information automatically.
     obj.key=eqtn; % handle will push this information automatically.
